Add filter-based Azure native subscription and resource group lookups
Introduce NativeSubscriptionsByFilter and NativeResourceGroupsByFilter,
which accept structured filter criteria (effective SLA, name substring,
region, subscription) instead of the substring-only and subscription-ID
lookups they replace. The old functions are deprecated. Also add
WaitForNativeSubscription so AddSubscription can block until a protection
feature's native subscription is available.

Split the growing core package into deployment, feature, hierarchy,
pagination, and task_chain files, and register the Azure native resource
group and cloud native tag rule inventory object types.
