+++ pkg/polaris/azure/native_account.go
@@ -21,8 +21,12 @@
 package azure
 
 import (
+	"cmp"
 	"context"
+	"errors"
 	"fmt"
+	"slices"
+	"time"
 
 	"github.com/google/uuid"
 	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/graphql"
@@ -31,21 +35,21 @@ import (
 )
 
 // NativeSubscriptionByID returns the native subscription with the specified RSC
-// native cloud account ID.
-func (a API) NativeSubscriptionByID(ctx context.Context, nativeCloudAccountID uuid.UUID) (azure.NativeSubscription, error) {
+// object ID.
+func (a API) NativeSubscriptionByID(ctx context.Context, objectID uuid.UUID) (azure.NativeSubscription, error) {
 	a.log.Print(log.Trace)
 
-	natives, err := a.NativeSubscriptions(ctx, "")
+	natives, err := a.NativeSubscriptionsByFilter(ctx, azure.NativeSubscriptionFilters{})
 	if err != nil {
 		return azure.NativeSubscription{}, fmt.Errorf("failed to list native subscriptions: %s", err)
 	}
 	for _, native := range natives {
-		if native.ID == nativeCloudAccountID {
+		if native.ID == objectID {
 			return native, nil
 		}
 	}
 
-	return azure.NativeSubscription{}, fmt.Errorf("native subscription %q %w", nativeCloudAccountID, graphql.ErrNotFound)
+	return azure.NativeSubscription{}, fmt.Errorf("native subscription %q %w", objectID, graphql.ErrNotFound)
 }
 
 // NativeSubscriptionByCloudAccountID returns the native subscription with the
@@ -53,7 +57,7 @@ func (a API) NativeSubscriptionByID(ctx context.Context, nativeCloudAccountID uu
 func (a API) NativeSubscriptionByCloudAccountID(ctx context.Context, cloudAccountID uuid.UUID) (azure.NativeSubscription, error) {
 	a.log.Print(log.Trace)
 
-	natives, err := a.NativeSubscriptions(ctx, "")
+	natives, err := a.NativeSubscriptionsByFilter(ctx, azure.NativeSubscriptionFilters{})
 	if err != nil {
 		return azure.NativeSubscription{}, fmt.Errorf("failed to list native subscriptions: %s", err)
 	}
@@ -63,23 +67,74 @@ func (a API) NativeSubscriptionByCloudAccountID(ctx context.Context, cloudAccoun
 		}
 	}
 
-	return azure.NativeSubscription{}, fmt.Errorf("native subscription for subscription %q %w", cloudAccountID, graphql.ErrNotFound)
+	return azure.NativeSubscription{}, fmt.Errorf("native subscription with cloud account id %q %w", cloudAccountID, graphql.ErrNotFound)
 }
 
-// NativeSubscriptions returns all native subscriptions matching the specified
-// filter. The filter can be used to search for a substring in the subscription
-// name.
+// Deprecated: use NativeSubscriptionsByFilter instead.
 func (a API) NativeSubscriptions(ctx context.Context, filter string) ([]azure.NativeSubscription, error) {
 	return azure.Wrap(a.client).NativeSubscriptions(ctx, filter)
 }
 
-// NativeResourceGroups returns Azure resource groups visible to RSC,
-// optionally filtered to the given subscription IDs and to those whose name
-// contains nameSubstring. A nil/empty subscription slice returns resource
-// groups across all managed subscriptions; the RSC API requires a non-empty
-// subscription filter, so in that case the list of managed subscriptions is
-// fetched and used as the filter. Passing an empty nameSubstring disables the
-// name filter.
+// NativeSubscriptionsByFilter returns all native subscriptions matching the
+// specified filters. See azure.NativeSubscriptionFilters for the available
+// filter criteria; a zero value returns all native subscriptions. Results are
+// sorted by name in ascending order.
+func (a API) NativeSubscriptionsByFilter(ctx context.Context, filters azure.NativeSubscriptionFilters) ([]azure.NativeSubscription, error) {
+	nativeSubs, err := azure.Wrap(a.client).NativeSubscriptionsByFilter(ctx, filters)
+	if err != nil {
+		return nil, fmt.Errorf("failed to list native subscriptions: %s", err)
+	}
+
+	// If an RSC subscription refresh runs at the same time as the subscriptions
+	// are read, the result might contain duplicates, remove those.
+	slices.SortFunc(nativeSubs, func(lhs, rhs azure.NativeSubscription) int {
+		return cmp.Compare(lhs.ID.String(), rhs.ID.String())
+	})
+	nativeSubs = slices.CompactFunc(nativeSubs, func(lhs, rhs azure.NativeSubscription) bool {
+		return lhs.ID == rhs.ID
+	})
+
+	slices.SortFunc(nativeSubs, func(lhs, rhs azure.NativeSubscription) int {
+		return cmp.Compare(lhs.Name, rhs.Name)
+	})
+
+	return nativeSubs, nil
+}
+
+// WaitForNativeSubscription blocks until a native subscription for the
+// specified RSC cloud account ID becomes available, polling every 30 seconds.
+// It returns nil once the native subscription is found, or an error if the
+// context is cancelled, the timeout is exceeded, or the lookup fails for any
+// reason other than the subscription not yet existing.
+func (a API) WaitForNativeSubscription(ctx context.Context, cloudAccountID uuid.UUID) error {
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Minute)
+	defer cancel()
+
+	a.log.Printf(log.Debug, "Waiting for native subscription with cloud account ID %q to become available", cloudAccountID)
+	for {
+		_, err := a.NativeSubscriptionByCloudAccountID(ctx, cloudAccountID)
+		switch {
+		case err == nil:
+			return nil
+		case errors.Is(err, context.DeadlineExceeded):
+			return fmt.Errorf("%w while waiting for native subscription with cloud account id %q to become available",
+				context.DeadlineExceeded, cloudAccountID)
+		case !errors.Is(err, graphql.ErrNotFound):
+			return err
+		}
+
+		a.log.Printf(log.Debug, "Native subscription with cloud account ID %q still not available", cloudAccountID)
+
+		select {
+		case <-ctx.Done():
+			return fmt.Errorf("%w while waiting for native subscription with cloud account id %q to become available",
+				context.DeadlineExceeded, cloudAccountID)
+		case <-time.After(30 * time.Second):
+		}
+	}
+}
+
+// Deprecated: use NativeResourceGroupsByFilter instead.
 func (a API) NativeResourceGroups(ctx context.Context, subscriptionIDs []uuid.UUID, nameSubstring string) ([]azure.NativeResourceGroup, error) {
 	if len(subscriptionIDs) == 0 {
 		subs, err := a.NativeSubscriptions(ctx, "")
@@ -95,4 +150,33 @@ func (a API) NativeResourceGroups(ctx context.Context, subscriptionIDs []uuid.UU
 		}
 	}
 	return azure.Wrap(a.client).NativeResourceGroups(ctx, subscriptionIDs, nameSubstring)
+}
+
+// NativeResourceGroupsByFilter returns all Azure resource groups matching the
+// specified filters. See azure.ResourceGroupFilters for the available filter
+// criteria; a zero value returns all resource groups. Results are sorted by
+// subscription name, then by resource group name.
+func (a API) NativeResourceGroupsByFilter(ctx context.Context, filters azure.ResourceGroupFilters) ([]azure.NativeResourceGroup, error) {
+	groups, err := azure.Wrap(a.client).NativeResourceGroupsByFilter(ctx, filters)
+	if err != nil {
+		return nil, err
+	}
+
+	// If an RSC subscription refresh run at the same time as the resource
+	// groups are read, the result might contain duplicates, remove those.
+	slices.SortFunc(groups, func(lhs, rhs azure.NativeResourceGroup) int {
+		return cmp.Compare(lhs.ID, rhs.ID)
+	})
+	groups = slices.CompactFunc(groups, func(lhs, rhs azure.NativeResourceGroup) bool {
+		return lhs.ID == rhs.ID
+	})
+
+	slices.SortFunc(groups, func(lhs, rhs azure.NativeResourceGroup) int {
+		return cmp.Or(
+			cmp.Compare(lhs.Subscription.Name, rhs.Subscription.Name),
+			cmp.Compare(lhs.Name, rhs.Name),
+		)
+	})
+
+	return groups, nil
 }
