+++ pkg/polaris/graphql/core/core.go
@@ -21,25 +21,30 @@
 // DEALINGS IN THE SOFTWARE.
 
 // Package core provides a low-level interface to core GraphQL queries provided
-// by the Polaris platform. E.g., task chains and enum definitions.
+// by the RSC platform. E.g., task chains and enum definitions.
 package core
 
 import (
-	"cmp"
 	"context"
 	"encoding/json"
-	"errors"
-	"fmt"
-	"slices"
 	"strings"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/graphql"
-	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/graphql/hierarchy"
-	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/graphql/sla"
 	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/log"
 )
+
+// API wraps around GraphQL clients to give them the RSC Core API.
+type API struct {
+	Version string // Deprecated: use GQL.DeploymentVersion
+	GQL     *graphql.Client
+	log     log.Logger
+}
+
+// Wrap the GraphQL client in the Core API.
+func Wrap(gql *graphql.Client) API {
+	return API{GQL: gql, log: gql.Log()}
+}
 
 type CloudVendor string
 
@@ -62,332 +67,7 @@ const (
 	UpdateRegions       CloudAccountAction = "UPDATE_REGIONS"
 )
 
-// PageInfo describes the result page metadata returned alongside a Relay-style
-// cursor connection.
-type PageInfo struct {
-	StartCursor     string `json:"startCursor"`
-	EndCursor       string `json:"endCursor"`
-	HasPreviousPage bool   `json:"hasPreviousPage"`
-	HasNextPage     bool   `json:"hasNextPage"`
-}
-
-// PermissionGroup represents a named set of permissions for a feature. Note,
-// not all permission groups are applicable to all features.
-type PermissionGroup string
-
-// Note: When adding new PermissionGroup constants, also add them to the
-// canaryPermissionGroups in permission_group_canary_test.go.
-const (
-	PermissionGroupAKSCustomPrivateDNSZone       PermissionGroup = "AKS_CUSTOM_PRIVATE_DNS_ZONE"
-	PermissionGroupAlloyDB                       PermissionGroup = "ALLOYDB"
-	PermissionGroupAutomatedNetworkingSetup      PermissionGroup = "AUTOMATED_NETWORKING_SETUP"
-	PermissionGroupBaaSBasic                     PermissionGroup = "BAAS_BASIC"
-	PermissionGroupBackupV2                      PermissionGroup = "BACKUP_V2"
-	PermissionGroupBasic                         PermissionGroup = "BASIC"
-	PermissionGroupCCES                          PermissionGroup = "CLOUD_CLUSTER_ES"
-	PermissionGroupCloudSQL                      PermissionGroup = "CLOUDSQL"
-	PermissionGroupCustomerHostedLogging         PermissionGroup = "CUSTOMER_HOSTED_LOGGING"
-	PermissionGroupCustomerManagedCluster        PermissionGroup = "CUSTOMER_MANAGED_BASIC"
-	PermissionGroupCustomerManagedStorageIndexng PermissionGroup = "CUSTOMER_MANAGED_STORAGE_INDEXING"
-	PermissionGroupDataCenterConsolidation       PermissionGroup = "DATA_CENTER_CONSOLIDATION"
-	PermissionGroupDataCenterImmutability        PermissionGroup = "DATA_CENTER_IMMUTABILITY"
-	PermissionGroupDataCenterKMS                 PermissionGroup = "DATA_CENTER_KMS"
-	PermissionGroupDownloadFile                  PermissionGroup = "DOWNLOAD_FILE"
-	PermissionGroupEncryption                    PermissionGroup = "ENCRYPTION"
-	PermissionGroupExportAndRestore              PermissionGroup = "EXPORT_AND_RESTORE"
-	PermissionGroupExportAndRestorePowerOffVM    PermissionGroup = "EXPORT_AND_RESTORE_POWER_OFF_VM"
-	PermissionGroupExportPowerOff                PermissionGroup = "EXPORT_POWER_OFF"
-	PermissionGroupExportPowerOn                 PermissionGroup = "EXPORT_POWER_ON"
-	PermissionGroupFileLevelRecovery             PermissionGroup = "FILE_LEVEL_RECOVERY"
-	PermissionGroupInvalid                       PermissionGroup = "GROUP_UNSPECIFIED"
-	PermissionGroupKMSKeySharing                 PermissionGroup = "KMS_KEY_SHARING"
-	PermissionGroupNATGateway                    PermissionGroup = "NAT_GATEWAY"
-	PermissionGroupPrivateEndpoints              PermissionGroup = "PRIVATE_ENDPOINTS"
-	PermissionGroupRecovery                      PermissionGroup = "RECOVERY"
-	PermissionGroupRecoveryNetworking            PermissionGroup = "RECOVERY_NETWORKING"
-	PermissionGroupRestore                       PermissionGroup = "RESTORE"
-	PermissionGroupRSCManagedCluster             PermissionGroup = "RSC_MANAGED_CLUSTER"
-	PermissionGroupSAPHanaSSBasic                PermissionGroup = "SAP_HANA_SS_BASIC"
-	PermissionGroupSAPHanaSSRecovery             PermissionGroup = "SAP_HANA_SS_RECOVERY"
-	PermissionGroupServiceEndpointAutomation     PermissionGroup = "SERVICE_ENDPOINT_AUTOMATION"
-	PermissionGroupSnapshotPrivateAccess         PermissionGroup = "SNAPSHOT_PRIVATE_ACCESS"
-	PermissionGroupSQLArchival                   PermissionGroup = "SQL_ARCHIVAL"
-)
-
-// SortOrder represents the valid sort order values.
-type SortOrder string
-
-const (
-	SortOrderAsc  SortOrder = "ASC"
-	SortOrderDesc SortOrder = "DESC"
-)
-
-// Pagination represents pagination parameters.
-type Pagination struct {
-	First     *int
-	After     *string
-	Last      *int
-	Before    *string
-	SortOrder *SortOrder
-}
-
-// Feature represents a Polaris cloud account feature with a set of permission
-// groups. If the PermissionGroups field is nil then the full set of permissions
-// are used for the feature.
-type Feature struct {
-	Name             string            `json:"featureType"`
-	PermissionGroups []PermissionGroup `json:"permissionsGroups"`
-}
-
-// Equal returns true if the features have the same name. Note, this function
-// does not compare the permission groups.
-func (feature Feature) Equal(other Feature) bool {
-	return feature.Name == other.Name
-}
-
-// DeepEqual returns true if the features are equal. The features are equal if
-// they have the same name and the same permission groups.
-func (feature Feature) DeepEqual(other Feature) bool {
-	if !feature.Equal(other) {
-		return false
-	}
-
-	set := make(map[PermissionGroup]struct{}, len(feature.PermissionGroups))
-	for _, permissionGroup := range feature.PermissionGroups {
-		set[permissionGroup] = struct{}{}
-	}
-	for _, permissionGroup := range other.PermissionGroups {
-		if _, ok := set[permissionGroup]; !ok {
-			return false
-		}
-		delete(set, permissionGroup)
-	}
-
-	return len(set) == 0
-}
-
-// HasPermissionGroup returns true if the feature has the specified permission
-// group.
-func (feature Feature) HasPermissionGroup(permissionGroup PermissionGroup) bool {
-	return slices.Contains(feature.PermissionGroups, permissionGroup)
-}
-
-// String returns a string representation of the feature.
-func (feature Feature) String() string {
-	if len(feature.PermissionGroups) == 0 {
-		return feature.Name
-	}
-
-	var buf strings.Builder
-	permissionGroups := slices.Clone(feature.PermissionGroups)
-	slices.Sort(permissionGroups)
-	for _, permissionGroup := range permissionGroups {
-		buf.WriteString(string(permissionGroup))
-		buf.WriteString(",")
-	}
-
-	return fmt.Sprintf("%s(%s)", feature.Name, buf.String()[:buf.Len()-1])
-}
-
-// WithPermissionGroups returns a copy of the feature with the specified
-// permission groups added.
-func (feature Feature) WithPermissionGroups(permissionGroups ...PermissionGroup) Feature {
-	groups := append(feature.PermissionGroups, permissionGroups...)
-	return Feature{Name: feature.Name, PermissionGroups: groups}
-}
-
-var (
-	FeatureInvalid                                     = Feature{Name: ""}
-	FeatureAll                                         = Feature{Name: "ALL"}
-	FeatureAppFlows                                    = Feature{Name: "APP_FLOWS"}
-	FeatureArchival                                    = Feature{Name: "ARCHIVAL"}
-	FeatureAzureDevOpsDeveloperCollaborationProtection = Feature{Name: "AZURE_DEVOPS_DEVELOPER_COLLABORATION_PROTECTION"}
-	FeatureAzureDevOpsProtection                       = Feature{Name: "AZURE_DEVOPS_PROTECTION"} // Deprecated: use FeatureAzureDevOpsRepositoryProtection
-	FeatureAzureDevOpsRepositoryProtection             = Feature{Name: "AZURE_DEVOPS_REPOSITORY_PROTECTION"}
-	FeatureAzureSQLDBProtection                        = Feature{Name: "AZURE_SQL_DB_PROTECTION"}
-	FeatureAzureSQLMIProtection                        = Feature{Name: "AZURE_SQL_MI_PROTECTION"}
-	FeatureCloudAccounts                               = Feature{Name: "CLOUDACCOUNTS"} // Deprecated: no replacement.
-	FeatureCloudDiscovery                              = Feature{Name: "CLOUD_DISCOVERY"}
-	FeatureCloudNativeArchival                         = Feature{Name: "CLOUD_NATIVE_ARCHIVAL"}
-	FeatureCloudNativeArchivalEncryption               = Feature{Name: "CLOUD_NATIVE_ARCHIVAL_ENCRYPTION"}
-	FeatureCloudNativeBlobProtection                   = Feature{Name: "CLOUD_NATIVE_BLOB_PROTECTION"}
-	FeatureCloudNativeDynamoDBProtection               = Feature{Name: "CLOUD_NATIVE_DYNAMODB_PROTECTION"}
-	FeatureCloudNativeProtection                       = Feature{Name: "CLOUD_NATIVE_PROTECTION"}
-	FeatureCloudNativeS3Protection                     = Feature{Name: "CLOUD_NATIVE_S3_PROTECTION"}
-	FeatureCyberRecoveryDataClassificationData         = Feature{Name: "CYBERRECOVERY_DATA_CLASSIFICATION_DATA"}
-	FeatureCyberRecoveryDataClassificationMetadata     = Feature{Name: "CYBERRECOVERY_DATA_CLASSIFICATION_METADATA"}
-	FeatureDSPMData                                    = Feature{Name: "DSPM_DATA"}
-	FeatureDSPMMetadata                                = Feature{Name: "DSPM_METADATA"}
-	FeatureExocompute                                  = Feature{Name: "EXOCOMPUTE"}
-	FeatureGCPSharedVPCHost                            = Feature{Name: "GCP_SHARED_VPC_HOST"}
-	FeatureGitHubRepositoryProtection                  = Feature{Name: "GITHUB_REPOSITORY_PROTECTION"}
-	FeatureKubernetesProtection                        = Feature{Name: "KUBERNETES_PROTECTION"}
-	FeatureLaminarCrossAccount                         = Feature{Name: "LAMINAR_CROSS_ACCOUNT"}
-	FeatureLaminarInternal                             = Feature{Name: "LAMINAR_INTERNAL"}
-	FeatureLaminarOutpostApplication                   = Feature{Name: "LAMINAR_OUTPOST_APPLICATION"}
-	FeatureLaminarOutpostManagedIdentity               = Feature{Name: "LAMINAR_OUTPOST_MANAGED_IDENTITY"}
-	FeatureLaminarTargetApplication                    = Feature{Name: "LAMINAR_TARGET_APPLICATION"}
-	FeatureLaminarTargetManagedIdentity                = Feature{Name: "LAMINAR_TARGET_MANAGED_IDENTITY"}
-	FeatureOutpost                                     = Feature{Name: "OUTPOST"}
-	FeatureRDSProtection                               = Feature{Name: "RDS_PROTECTION"}
-	FeatureRoleChaining                                = Feature{Name: "ROLE_CHAINING"}
-	FeatureServerAndApps                               = Feature{Name: "SERVERS_AND_APPS"}
-)
-
-var validFeatures = map[string]struct{}{
-	FeatureAll.Name:      {},
-	FeatureAppFlows.Name: {},
-	FeatureArchival.Name: {},
-	FeatureAzureDevOpsDeveloperCollaborationProtection.Name: {},
-	FeatureAzureDevOpsProtection.Name:                       {}, // Deprecated: use FeatureAzureDevOpsRepositoryProtection
-	FeatureAzureDevOpsRepositoryProtection.Name:             {},
-	FeatureAzureSQLDBProtection.Name:                        {},
-	FeatureAzureSQLMIProtection.Name:                        {},
-	FeatureCloudAccounts.Name:                               {}, // Deprecated: no replacement.
-	FeatureCloudDiscovery.Name:                              {},
-	FeatureCloudNativeArchival.Name:                         {},
-	FeatureCloudNativeArchivalEncryption.Name:               {},
-	FeatureCloudNativeBlobProtection.Name:                   {},
-	FeatureCloudNativeDynamoDBProtection.Name:               {},
-	FeatureCloudNativeProtection.Name:                       {},
-	FeatureCloudNativeS3Protection.Name:                     {},
-	FeatureCyberRecoveryDataClassificationData.Name:         {},
-	FeatureCyberRecoveryDataClassificationMetadata.Name:     {},
-	FeatureDSPMData.Name:                                    {},
-	FeatureDSPMMetadata.Name:                                {},
-	FeatureExocompute.Name:                                  {},
-	FeatureGCPSharedVPCHost.Name:                            {},
-	FeatureGitHubRepositoryProtection.Name:                  {},
-	FeatureKubernetesProtection.Name:                        {},
-	FeatureLaminarCrossAccount.Name:                         {},
-	FeatureLaminarInternal.Name:                             {},
-	FeatureLaminarOutpostApplication.Name:                   {},
-	FeatureLaminarOutpostManagedIdentity.Name:               {},
-	FeatureLaminarTargetApplication.Name:                    {},
-	FeatureLaminarTargetManagedIdentity.Name:                {},
-	FeatureOutpost.Name:                                     {},
-	FeatureRDSProtection.Name:                               {},
-	FeatureRoleChaining.Name:                                {},
-	FeatureServerAndApps.Name:                               {},
-}
-
-// AllProtectionFeatures returns the protection features for the specified cloud
-// vendor.
-func AllProtectionFeatures(cloud CloudVendor) []Feature {
-	switch cloud {
-	case CloudVendorAWS:
-		return []Feature{
-			FeatureCloudNativeDynamoDBProtection,
-			FeatureCloudNativeProtection,
-			FeatureCloudNativeS3Protection,
-			FeatureKubernetesProtection,
-			FeatureRDSProtection,
-		}
-	case CloudVendorAzure:
-		return []Feature{
-			FeatureAzureSQLDBProtection,
-			FeatureAzureSQLMIProtection,
-			FeatureCloudNativeBlobProtection,
-			FeatureCloudNativeProtection,
-			FeatureKubernetesProtection,
-		}
-	case CloudVendorGCP:
-		return []Feature{
-			FeatureCloudNativeProtection,
-		}
-	default:
-		return nil
-	}
-}
-
-// FeatureNames returns the names of the features.
-func FeatureNames(features []Feature) []string {
-	var names []string
-	for _, feature := range features {
-		names = append(names, feature.Name)
-	}
-
-	return names
-}
-
-// LookupFeature returns the specified feature if it exists in the feature
-// slice.
-func LookupFeature(features []Feature, feature Feature) (Feature, bool) {
-	for _, f := range features {
-		if f.Equal(feature) {
-			return f, true
-		}
-	}
-
-	return Feature{}, false
-}
-
-// ValidateRoleChaining returns an error if ROLE_CHAINING is combined with
-// other features. The ROLE_CHAINING feature is mutually exclusive with all
-// other features.
-func ValidateRoleChaining(features []Feature) error {
-	if len(features) < 2 {
-		return nil
-	}
-	if _, ok := LookupFeature(features, FeatureRoleChaining); ok {
-		return errors.New("ROLE_CHAINING is mutually exclusive with all other features")
-	}
-	return nil
-}
-
-// FilterFeaturesOnPermissionGroups verifies that all features either have no
-// permission groups or all have permission groups. The features are returned
-// in two different slices, depending on whether they have permission groups
-// or not.
-func FilterFeaturesOnPermissionGroups(features []Feature) ([]string, []Feature, error) {
-	if len(features) == 0 {
-		return nil, nil, errors.New("no features specified")
-	}
-
-	// Check that all features have the same use of permission groups.
-	usePG := len(features[0].PermissionGroups) > 0
-	for _, feature := range features[1:] {
-		if pg := len(feature.PermissionGroups) > 0; pg != usePG {
-			return nil, nil, errors.New("features with and without permission groups cannot be mixed")
-		}
-	}
-	if usePG {
-		return nil, features, nil
-	}
-
-	return FeatureNames(features), nil, nil
-}
-
-// Deprecated: use Feature.Name instead.
-func FormatFeature(feature Feature) string {
-	return strings.ReplaceAll(strings.ToLower(feature.Name), "_", "-")
-}
-
-// Deprecated: use Feature{Name: <feature>} instead or ParseFeatureNoValidation
-// if you need to remain backwards compatible with previously accepted feature
-// names.
-func ParseFeature(feature string) (Feature, error) {
-	f := ParseFeatureNoValidation(feature)
-	if _, ok := validFeatures[f.Name]; ok {
-		return f, nil
-	}
-
-	return FeatureInvalid, fmt.Errorf("invalid feature: %s", feature)
-}
-
-// ParseFeatureNoValidation returns the Feature matching the given feature name.
-// No validation is performed.
-func ParseFeatureNoValidation(feature string) Feature {
-	return Feature{Name: strings.ToUpper(strings.ReplaceAll(feature, "-", "_"))}
-}
-
-const (
-	// The number of attempts before failing to wait for the Korg job when the
-	// error returned is a 403, objects not authorized.
-	waitAttempts = 50
-)
-
-// Status represents a Polaris cloud account status.
+// Status represents an RSC cloud account status.
 type Status string
 
 const (
@@ -404,271 +84,10 @@ func FormatStatus(status Status) string {
 	return strings.ReplaceAll(strings.ToLower(string(status)), "_", "-")
 }
 
-// TaskChainState represents the state of a Polaris task chain.
-type TaskChainState string
-
-const (
-	TaskChainInvalid   TaskChainState = ""
-	TaskChainCanceled  TaskChainState = "CANCELED"
-	TaskChainCanceling TaskChainState = "CANCELING"
-	TaskChainFailed    TaskChainState = "FAILED"
-	TaskChainReady     TaskChainState = "READY"
-	TaskChainRunning   TaskChainState = "RUNNING"
-	TaskChainSucceeded TaskChainState = "SUCCEEDED"
-	TaskChainUndoing   TaskChainState = "UNDOING"
-)
-
-// API wraps around GraphQL clients to give them the Polaris Core API.
-type API struct {
-	Version string // Deprecated: use GQL.DeploymentVersion
-	GQL     *graphql.Client
-	log     log.Logger
-}
-
-// Wrap the GraphQL client in the Core API.
-func Wrap(gql *graphql.Client) API {
-	return API{GQL: gql, log: gql.Log()}
-}
-
-// TaskChain is a collection of sequential tasks that all must complete for the
-// task chain to be considered complete.
-type TaskChain struct {
-	ID          int64          `json:"id"`
-	TaskChainID uuid.UUID      `json:"taskchainUuid"`
-	State       TaskChainState `json:"state"`
-}
-
-// KorgTaskChainStatus returns the task chain for the specified task chain id.
-// If the task chain id refers to a task chain that was just created, its state
-// might not have reached ready yet. This can be detected by state being
-// TaskChainInvalid and error is nil.
-func (a API) KorgTaskChainStatus(ctx context.Context, taskChainID uuid.UUID) (TaskChain, error) {
-	a.log.Print(log.Trace)
-
-	query := getKorgTaskchainStatusQuery
-	buf, err := a.GQL.Request(ctx, query, struct {
-		TaskChainID uuid.UUID `json:"taskchainId,omitempty"`
-	}{TaskChainID: taskChainID})
-	if err != nil {
-		return TaskChain{}, graphql.RequestError(query, err)
-	}
-
-	var payload struct {
-		Data struct {
-			Query struct {
-				TaskChain TaskChain `json:"taskchain"`
-			} `json:"getKorgTaskchainStatus"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal(buf, &payload); err != nil {
-		return TaskChain{}, graphql.UnmarshalError(query, err)
-	}
-
-	return payload.Data.Query.TaskChain, nil
-}
-
-// WaitForTaskChain blocks until the Polaris task chain with the specified task
-// chain id has completed. When the task chain completes, the final state of the
-// task chain is returned. The wait parameter specifies the amount of time to
-// wait before requesting another task status update.
-func (a API) WaitForTaskChain(ctx context.Context, taskChainID uuid.UUID, wait time.Duration) (TaskChainState, error) {
-	a.log.Print(log.Trace)
-
-	attempt := 0
-	for {
-		taskChain, err := a.KorgTaskChainStatus(ctx, taskChainID)
-		if err != nil {
-			var gqlErr graphql.GQLError
-			if !errors.As(err, &gqlErr) || len(gqlErr.Errors) < 1 {
-				return TaskChainInvalid, fmt.Errorf("failed to get tashchain status for %s: %s", taskChainID, err)
-			}
-			for _, e := range gqlErr.Errors {
-				if e.Extensions.Code == 403 || e.Extensions.Code == 500 {
-					continue // Could be a RBAC error that eventually goes away, keep retrying.
-				}
-				return TaskChainInvalid, fmt.Errorf("unexpected error code when getting tashchain status for %s: %v", taskChainID, err)
-			}
-			if attempt++; attempt > waitAttempts {
-				return TaskChainInvalid, fmt.Errorf("failed to get tashchain status for %s after %d attempts: %s", taskChainID, attempt, err)
-			}
-			a.log.Printf(log.Debug, "RBAC not ready (attempt: %d)", attempt)
-		}
-
-		if taskChain.State == TaskChainSucceeded || taskChain.State == TaskChainCanceled || taskChain.State == TaskChainFailed {
-			return taskChain.State, nil
-		}
-
-		a.log.Printf(log.Debug, "Waiting for Polaris task chain: %s", taskChainID)
-
-		select {
-		case <-time.After(wait):
-		case <-ctx.Done():
-			return TaskChainInvalid, ctx.Err()
-		}
-	}
-}
-
-// WaitForFeatureDisableTaskChain waits for the feature disable task chain to
-// finish. If an error occurs while waiting for the task chain or the task chain
-// ends in a failed state, an error is returned.
-func (a API) WaitForFeatureDisableTaskChain(ctx context.Context, taskChainID uuid.UUID, featureStatus func(ctx context.Context) (bool, error)) error {
-	a.log.Print(log.Trace)
-
-	for {
-		// Check the status of the task chain.
-		taskChain, err := a.KorgTaskChainStatus(ctx, taskChainID)
-		if err != nil {
-			var gqlErr graphql.GQLError
-			if !errors.As(err, &gqlErr) || len(gqlErr.Errors) < 1 {
-				return fmt.Errorf("failed to retrieve taskchain status: %s", err)
-			}
-			for _, e := range gqlErr.Errors {
-				if e.Extensions.Code == 403 || e.Extensions.Code == 500 {
-					continue // Could be a RBAC error that eventually goes away, keep retrying.
-				}
-				return fmt.Errorf("unexpected error code when getting tashchain status for %s: %v", taskChainID, err)
-			}
-
-			// If the task chain RBAC is not yet ready, we fall back to checking
-			// the status of the account feature.
-			if disabled, err := featureStatus(ctx); disabled || err != nil {
-				return err
-			}
-
-			a.log.Printf(log.Debug, "Task chain RBAC not ready")
-		} else {
-			if taskChain.State == TaskChainSucceeded {
-				return nil
-			}
-			if taskChain.State == TaskChainCanceled || taskChain.State == TaskChainFailed {
-				return fmt.Errorf("taskchain failed: task chain state is %s", taskChain.State)
-			}
-		}
-
-		a.log.Printf(log.Debug, "Waiting for task chain: %s", taskChainID)
-		select {
-		case <-time.After(10 * time.Second):
-		case <-ctx.Done():
-			return ctx.Err()
-		}
-	}
-}
-
-// Deprecated: use GQL.DeploymentVersion.
-func (a API) DeploymentVersion(ctx context.Context) (string, error) {
-	a.log.Print(log.Trace)
-
-	query := deploymentVersionQuery
-	buf, err := a.GQL.Request(ctx, query, struct{}{})
-	if err != nil {
-		return "", graphql.RequestError(query, err)
-	}
-
-	var payload struct {
-		Data struct {
-			DeploymentVersion string `json:"deploymentVersion"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal(buf, &payload); err != nil {
-		return "", graphql.UnmarshalError(query, err)
-	}
-
-	return payload.Data.DeploymentVersion, nil
-}
-
-// DeploymentIPAddresses returns the deployment IP addresses.
-func (a API) DeploymentIPAddresses(ctx context.Context) ([]string, error) {
-	a.log.Print(log.Trace)
-
-	query := allDeploymentIpAddressesQuery
-	buf, err := a.GQL.Request(ctx, query, struct{}{})
-	if err != nil {
-		return nil, graphql.RequestError(query, err)
-	}
-
-	var payload struct {
-		Data struct {
-			DeploymentIPAddresses []string `json:"allDeploymentIpAddresses"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal(buf, &payload); err != nil {
-		return nil, graphql.UnmarshalError(query, err)
-	}
-
-	return payload.Data.DeploymentIPAddresses, nil
-}
-
-// EnabledFeaturesForAccount returns all features enable for the RSC account.
-func (a API) EnabledFeaturesForAccount(ctx context.Context) ([]Feature, error) {
-	a.log.Print(log.Trace)
-
-	query := allEnabledFeaturesForAccountQuery
-	buf, err := a.GQL.Request(ctx, query, struct{}{})
-	if err != nil {
-		return nil, graphql.RequestError(query, err)
-	}
-
-	var payload struct {
-		Data struct {
-			Result struct {
-				Features []string `json:"features"`
-			} `json:"result"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal(buf, &payload); err != nil {
-		return nil, graphql.UnmarshalError(query, err)
-	}
-
-	var features []Feature
-	for _, feature := range payload.Data.Result.Features {
-		features = append(features, Feature{Name: feature})
-	}
-	slices.SortFunc(features, func(i, j Feature) int {
-		return cmp.Compare(i.Name, j.Name)
-	})
-
-	return features, nil
-}
-
-// FormatTimestamp converts a time.Time to RFC3339 format with milliseconds and Z suffix.
+// FormatTimestamp converts a time.Time to RFC3339 format with milliseconds and
+// Z suffix.
 func FormatTimestamp(t time.Time) string {
 	return t.UTC().Format("2006-01-02T15:04:05.000Z")
-}
-
-// HierarchyObjectType represents the type of a hierarchy object.
-//
-// Deprecated: Use hierarchy.ObjectType instead.
-type HierarchyObjectType = hierarchy.ObjectType
-
-// HierarchyObject represents an RSC hierarchy object with SLA information.
-//
-// Deprecated: Use sla.HierarchyObject instead.
-type HierarchyObject = sla.HierarchyObject
-
-// DoNotProtectSLAID is the special SLA domain ID used to indicate that an
-// object should not be protected. This is returned in configuredSlaDomain.ID
-// when "Do Not Protect" is directly assigned to an object.
-//
-// Deprecated: Use sla.DoNotProtectSLAID instead.
-const DoNotProtectSLAID = sla.DoNotProtectSLAID
-
-// UnprotectedSLAID is the special SLA domain ID used to indicate that an
-// object is unprotected (no SLA assigned). This is returned in
-// effectiveSlaDomain.ID when the object inherits no protection.
-//
-// Deprecated: Use sla.UnprotectedSLAID instead.
-const UnprotectedSLAID = sla.UnprotectedSLAID
-
-// HierarchyObjectByID returns the hierarchy object with the specified ID.
-// This can be used to query any hierarchy object (VMs, databases, tag rules,
-// etc.) and retrieve its SLA assignment information including the configured
-// and effective SLA domains.
-//
-// Deprecated: Use sla.ObjectByID instead.
-func (a API) HierarchyObjectByID(ctx context.Context, fid uuid.UUID) (HierarchyObject, error) {
-	a.log.Print(log.Trace)
-
-	return sla.ObjectByID(ctx, a.GQL, fid)
 }
 
 // ValuesByEnum returns the enum values for the specified enum name in the RSC GraphQL API.
